fix some render/redirect route stuff
The :who :where :what route placeholders were not working
because I had an unneccesary redirect in there in order
to make page requests "canonical" and I realised I didn't care
whether a request was "canonical" or not, so long as I
was able to find the relevant page.

Then I discovered I'd been doing needless work when I
served files, so I fixed that too.